Update GPUSampleMask from Long to UInt and improve mappings
The GPUSampleMask type has been refactored from Long to UInt, improving the efficiency of handling these values within the pipeline. This update required changes in RenderPipelineDescriptor.kt and adjustment of related usage instances. A new MappingTransformer, UIntToIntTransformer, has been added to manage this conversion. A slight log formatting correction was also made in CCodeGenerator.kt.

+++ wgpu4k/src/commonMain/kotlin/io.ygdrasil.wgpu/Types.kt
@@ -24,7 +24,7 @@ typealias GPUMapModeFlags = Number
 
 typealias GPUPipelineConstantValue = Number
 
-typealias GPUSampleMask = Long
+typealias GPUSampleMask = UInt
 
 typealias GPUShaderStageFlags = Number
